third_party/blink/web_tests/accessibility/aria-owns-presentational-node-edge-cases.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div role="presentation">
    <div id="parent"> <!-- Will get aria-owns="2" -->
    </div>
    <div id="2" role="presentation">
        <div id="3" role="button"></div>
        <div id="4" role="button"></div>
    </div>
</div>

<script>
test(function(t)
{
    const axParent = accessibilityController.accessibleElementById("parent");
    const axRoot = axParent.parentElement();
    assert_equals(axRoot.role, "AXRole: AXPresentational");
    assert_equals(axRoot.childrenCount, 2);

    document.getElementById('parent').setAttribute('aria-owns', '2');
    assert_equals(axParent.childrenCount, 1);
    assert_equals(axRoot.childrenCount, 1);  // Root no longer owns #2.

    const axChild = axParent.childAtIndex(0);
    assert_equals(axChild.role, "AXRole: AXPresentational");
}, "Aria-owns cannot be fooled by pointing it at a presentational node.");
</script>
